refactor: extract BackupRestoreView into modular sub-components
Refactored the 935-line BackupRestoreView component into a well-organized
component structure for better maintainability and reusability.

**Component Structure:**
- BackupRestoreView.tsx (102 lines) - Main orchestrator
- backup-restore/
  - StorageUsageDisplay.tsx (91 lines) - Storage quota monitoring
  - BackupSection.tsx (149 lines) - Backup creation with encryption
  - BackupFilePreview.tsx (244 lines) - File metadata & validation
  - CategorySelector.tsx (86 lines) - Category selection UI
  - RestoreSection.tsx (385 lines) - Restore orchestration
  - BackupHistoryList.tsx (42 lines) - Recent backups display
  - types.ts (66 lines) - Shared utilities and constants
  - index.ts (9 lines) - Barrel exports

**Benefits:**
- 89% reduction in main component size (935 → 102 lines)
- Clear separation of concerns
- Improved reusability (BackupFilePreview, CategorySelector)
- Easier to test and maintain
- All components < 400 lines
- No breaking changes

**Quality Assurance:**
- ✅ All 727 tests passing
- ✅ ESLint: 0 errors
- ✅ 100% backward compatibility
- ✅ Same functionality, better structure

Addresses PR #55 review suggestion for component extraction.
