+++ src/components/settings/backup-restore/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Barrel export for backup-restore components
+ */
+export { default as BackupSection } from './BackupSection';
+export { default as RestoreSection } from './RestoreSection';
+export { default as BackupFilePreview } from './BackupFilePreview';
+export { default as CategorySelector } from './CategorySelector';
+export { default as BackupHistoryList } from './BackupHistoryList';
+export { default as StorageUsageDisplay } from './StorageUsageDisplay';
+export * from './types';
